feat(*): add ai-agent to generate plot or made a query based on scraped data saved in csv file
